requirements.txt
boto3
awscli
streamlit
faiss-cpu
numpy
sentence-transformers